Move module static variables from World.cpp to the WorldImpl class.
Move room properties and the profile to the WorldImpl class.
Move object and timer arrays to the WorldImpl class.
Disable C4351, because we want to default-initialize arrays in the constructor initialization list.
Move the variables that keep track of the observed position of the player to the WorldImpl class.
Move the player variable to the WorldImpl class.